Move WithdrawFundsForm to model package.
