+++ src/main/java/io/bisq/api/service/v1/WalletResource.java
@@ -1,13 +1,10 @@
 package io.bisq.api.service.v1;
 
-import io.bisq.api.AmountTooLowException;
-import io.bisq.api.BisqProxy;
-import io.bisq.api.model.WalletAddress;
-import io.bisq.api.model.WalletAddressList;
-import io.bisq.api.model.WalletDetails;
-import io.bisq.api.model.WalletTransactionList;
 import bisq.core.btc.AddressEntryException;
 import bisq.core.btc.InsufficientFundsException;
+import io.bisq.api.AmountTooLowException;
+import io.bisq.api.BisqProxy;
+import io.bisq.api.model.*;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.Authorization;
